% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robots_validate.R
\name{print.robots_validations}
\alias{print.robots_validations}
\title{Print a robots.txt validation result}
\usage{
\method{print}{robots_validations}(x, ...)
}
\arguments{
\item{x}{A \code{robots_validations} object.}

\item{...}{Additional arguments, currently ignored.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a robots.txt validation result
}
